Move stacks to App.
