+++ App.js
@@ -2,10 +2,8 @@ import 'react-native-gesture-handler';
 import React from 'react';
 
 import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-
-import LibraryStackScreen from '@stacks/LibraryStack';
-import WishlistStackScreen from '@stacks/WishlistStack';
 
 import TabBarIcon from '@components/TabBarIcon';
 import { ROUTES } from '@constants/routes';
@@ -13,11 +11,53 @@ import { COLORS } from '@constants/colors';
 
 import starActive from '@assets/toolbar/ic_wishlist_active.png';
 import starInactive from '@assets/toolbar/ic_wishlist.png';
-
 import libraryActive from '@assets/toolbar/ic_library_active.png';
 import libraryInactive from '@assets/toolbar/ic_library.png';
 
+import LibraryScreen from '@screens/LibraryScreen';
+import BookDetailScreen from '@screens/BookDetailScreen';
+import WishlistScreen from '@screens/WishlistScreen';
+
+import { stackNavigatorConfig } from '@utils/stackConfig';
+
+const LibraryStack = createStackNavigator();
+const WishlistStack = createStackNavigator();
 const Tab = createBottomTabNavigator();
+
+function LibraryStackScreen() {
+  return (
+    <LibraryStack.Navigator
+      initialRouteName={ROUTES.Library.name}
+      headerMode="float"
+      screenOptions={stackNavigatorConfig}>
+      <LibraryStack.Screen
+        name={ROUTES.Library.name}
+        component={LibraryScreen}
+        options={{ title: ROUTES.Library.title }}
+      />
+      <LibraryStack.Screen
+        name={ROUTES.BookDetail.name}
+        component={BookDetailScreen}
+        options={{ title: ROUTES.BookDetail.title }}
+      />
+    </LibraryStack.Navigator>
+  );
+}
+
+function WishlistStackScreen() {
+  return (
+    <WishlistStack.Navigator
+      initialRouteName={ROUTES.Wishlist.name}
+      headerMode="float"
+      screenOptions={stackNavigatorConfig}>
+      <WishlistStack.Screen
+        name={ROUTES.Wishlist.name}
+        component={WishlistScreen}
+        options={{ title: ROUTES.Wishlist.title }}
+      />
+    </WishlistStack.Navigator>
+  );
+}
 
 export default function App() {
   return (
