+++ src/app/components/TabBarIcon/index.js
@@ -1,12 +1,8 @@
 import React from 'react';
 import { number, bool } from 'prop-types';
-import { Image, StyleSheet } from 'react-native';
+import { Image } from 'react-native';
 
-const styles = StyleSheet.create({
-  image: {
-    marginTop: 5
-  }
-});
+import styles from './styles';
 
 function TabBarIcon({ active, inactive, focused, size }) {
   return (
