+++ src/app/components/TabBarIcon/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  image: {
+    marginTop: 5
+  }
+});
+
+export default styles;
